+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [NEXT_RELEASE]
+### Changed
+- [server] Specific CPU and Memory limits for both deploy and release pods
+
 ## [0.5.0] - 2017-08-15
 ### Changed
 - Use gRPC instead of go-swagger